+++ book/guesser.go
@@ -80,9 +80,7 @@ var (
 		// <Title> / <SubTitle>
 		regexp.MustCompile(`^(?P<Title>.+)\s/\s(?P<SubTitle>.+)$`),
 		// <Title> (French Edition)
-		regexp.MustCompile(`^(?P<Title>.+)\s\p{Ps}(?i:.+?\sedition)\p{Pe}$`),
-		// <Title> (Littérature française)
-		regexp.MustCompile(`^(?P<Title>.+)\s\p{Ps}(?i:.*?littérature.+?)\p{Pe}$`),
+		regexp.MustCompile(`^(?P<Title>.+)\s\p{Ps}(?:.*(?i:[eé]dition|litt[eé]rature).*|\p{Lu}+|[0-9]{4})\p{Pe}$`),
 	}
 )